fix: fixed test leaving db item behind
