+++ tests/test_datasets_endpoint.py
@@ -2,15 +2,16 @@
 Dataset endpoint Lambda function tests. Working Data Lake AWS environment is
 required (run '$ cdk deploy' before running tests).
 """
-
+import json
 import logging
 import re
 
 import _pytest
+from mypy_boto3_lambda import LambdaClient
 from pytest import mark
 
 from backend.datasets import entrypoint
-from backend.utils import DATASET_TYPES
+from backend.utils import DATASET_TYPES, ResourceName
 
 from .utils import (
     Dataset,
@@ -382,3 +383,29 @@ def test_should_fail_if_deleting_not_existing_dataset(
             "message": f"Not Found: dataset '{dataset_id}' of type '{dataset_type}' does not exist"
         },
     }
+
+
+@mark.infrastructure
+def test_should_launch_datasets_endpoint_lambda_function(
+    lambda_client: LambdaClient,
+    datasets_db_teardown: _pytest.fixtures.FixtureDef[object],  # pylint:disable=unused-argument
+) -> None:
+    """
+    Test if datasets endpoint lambda can be successfully launched and has required permission to
+    create dataset in DB.
+    """
+
+    method = "POST"
+    body = {}
+    body["type"] = any_valid_dataset_type()
+    body["title"] = any_dataset_title()
+    body["owning_group"] = any_dataset_owning_group()
+
+    resp = lambda_client.invoke(
+        FunctionName=ResourceName.DATASETS_ENDPOINT_FUNCTION_NAME.value,
+        Payload=json.dumps({"httpMethod": method, "body": body}).encode(),
+        InvocationType="RequestResponse",
+    )
+    json_resp = json.load(resp["Payload"])
+
+    assert json_resp.get("statusCode") == 201, json_resp
